- 完成获取数据分析的接口（待完善）

+++ server/internal/pkg/domain/website_config.go
@@ -58,7 +58,7 @@ type WebSiteConfig struct {
 	// 分类数量
 	CategoryCount uint `bson:"category_count"`
 	// 访问量
-	ViewCount uint `bson:"view_count"`
+	ViewCount int64 `bson:"view_count"`
 	// 网站运行时间
 	LiveTime int64 `bson:"live_time"`
 	// 域名
